Multiple pairwise comparison tests with tidy data

Calculate parametric, non-parametric, robust, and Bayes Factor pairwise comparisons between group levels with corrections for multiple testing.

pairwise_comparisons(
  data,
  x,
  y,
  subject.id = NULL,
  type = "parametric",
  paired = FALSE,
  var.equal = FALSE,
  tr = 0.1,
  bf.prior = 0.707,
  p.adjust.method = "holm",
  k = 2L,
  ...
)

Arguments

data
A dataframe from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

subject.id
In case of repeated measures design (paired = TRUE, i.e.), this argument specifies the subject or repeated measures id. Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate.

type
Type of statistic expected ("parametric" or "nonparametric" or "robust" or "bayes").Corresponding abbreviations are also accepted: "p" (for parametric), "np" (nonparametric), "r" (robust), or "bf"resp.

paired
Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

var.equal
a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

tr
Trim level for the mean when carrying out robust tests. If you get error stating "Standard error cannot be computed because of Winsorized variance of 0 (e.g., due to ties). Try to decrease the trimming level.", try to play around with the value of tr, which is by default set to 0.1. Lowering the value might help.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

p.adjust.method
Adjustment method for p-values for multiple comparisons. Possible methods are: "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

...
Current ignored.

Value

A tibble dataframe containing two columns corresponding to group levels being compared with each other (group1 and group2) and p.value column corresponding to this comparison. The dataframe will also contain a p.value.label column containing a label for this p-value, in case this needs to be displayed in ggsignif::geom_ggsignif. In addition to these common columns across the different types of statistics, there will be additional columns specific to the type of test being run.

This function provides a unified syntax to carry out pairwise comparison tests and internally relies on other packages to carry out these tests. For more details about the included tests, see the documentation for the respective functions:

parametric : stats::pairwise.t.test() (paired) and PMCMRplus::gamesHowellTest() (unpaired)

non-parametric : PMCMRplus::durbinAllPairsTest() (paired) and PMCMRplus::kwAllPairsDunnTest() (unpaired)

robust : WRS2::rmmcp() (paired) and WRS2::lincon() (unpaired)

Bayes Factor : BayesFactor::ttestBF()

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(pairwiseComparisons)

# show me all columns and make the column titles bold
options(tibble.width = Inf, pillar.bold = TRUE, pillar.subtle_num = TRUE)

#------------------- between-subjects design ----------------------------

# parametric
# if `var.equal = TRUE`, then Student's t-test will be run
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "parametric",
  var.equal = TRUE,
  paired = FALSE,
  p.adjust.method = "none"
)

#> # A tibble: 3 x 6
#>   group1 group2     p.value test.details     p.value.adjustment
#>   <chr>  <chr>        <dbl> <chr>            <chr>             
#> 1 4      6      0.0106      Student's t-test None              
#> 2 4      8      0.000000207 Student's t-test None              
#> 3 6      8      0.00516     Student's t-test None              
#>   label                                  
#>   <chr>                                  
#> 1 list(~italic(p)[uncorrected]==0.011)   
#> 2 list(~italic(p)[uncorrected]==2.07e-07)
#> 3 list(~italic(p)[uncorrected]==0.005)   

# if `var.equal = FALSE`, then Games-Howell test will be run
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "parametric",
  var.equal = FALSE,
  paired = FALSE,
  p.adjust.method = "bonferroni"
)

#> # A tibble: 3 x 11
#>   group1 group2 statistic   p.value alternative method            distribution
#>   <chr>  <chr>      <dbl>     <dbl> <chr>       <chr>             <chr>       
#> 1 4      6           5.39 0.0125    two.sided   Games-Howell test q           
#> 2 4      8           9.11 0.0000124 two.sided   Games-Howell test q           
#> 3 6      8           5.12 0.0148    two.sided   Games-Howell test q           
#>   p.adjustment test.details      p.value.adjustment
#>   <chr>        <chr>             <chr>             
#> 1 none         Games-Howell test Bonferroni        
#> 2 none         Games-Howell test Bonferroni        
#> 3 none         Games-Howell test Bonferroni        
#>   label                                           
#>   <chr>                                           
#> 1 list(~italic(p)[Bonferroni-corrected]==0.012)   
#> 2 list(~italic(p)[Bonferroni-corrected]==1.24e-05)
#> 3 list(~italic(p)[Bonferroni-corrected]==0.015)   

# non-parametric (Dunn test)
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "nonparametric",
  paired = FALSE,
  p.adjust.method = "none"
)

#> # A tibble: 3 x 11
#>   group1 group2 statistic    p.value alternative method               
#>   <chr>  <chr>      <dbl>      <dbl> <chr>       <chr>                
#> 1 4      6           1.84 0.0663     two.sided   Dunn's all-pairs test
#> 2 4      8           4.76 0.00000198 two.sided   Dunn's all-pairs test
#> 3 6      8           2.22 0.0263     two.sided   Dunn's all-pairs test
#>   distribution p.adjustment test.details p.value.adjustment
#>   <chr>        <chr>        <chr>        <chr>             
#> 1 z            none         Dunn test    None              
#> 2 z            none         Dunn test    None              
#> 3 z            none         Dunn test    None              
#>   label                                  
#>   <chr>                                  
#> 1 list(~italic(p)[uncorrected]==0.066)   
#> 2 list(~italic(p)[uncorrected]==1.98e-06)
#> 3 list(~italic(p)[uncorrected]==0.026)   

# robust (Yuen's trimmed means t-test)
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "robust",
  paired = FALSE,
  p.adjust.method = "fdr"
)

#> # A tibble: 3 x 9
#>   group1 group2 estimate conf.low conf.high  p.value test.details             
#>   <chr>  <chr>     <dbl>    <dbl>     <dbl>    <dbl> <chr>                    
#> 1 4      6        -0.846    -1.51    -0.180 0.00625  Yuen's trimmed means test
#> 2 4      8        -1.68     -2.48    -0.878 0.000258 Yuen's trimmed means test
#> 3 6      8        -0.832    -1.53    -0.131 0.00625  Yuen's trimmed means test
#>   p.value.adjustment label                                    
#>   <chr>              <chr>                                    
#> 1 FDR                list(~italic(p)[FDR-corrected]==0.006)   
#> 2 FDR                list(~italic(p)[FDR-corrected]==2.58e-04)
#> 3 FDR                list(~italic(p)[FDR-corrected]==0.006)   

# Bayes Factor (Student's t-test)
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "bayes",
  paired = FALSE
)

#> # A tibble: 3 x 16
#>   group1 group2 term       estimate conf.low conf.high    pd rope.percentage
#>   <chr>  <chr>  <chr>         <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 4      6      Difference    0.686    0.252      1.10 0.992               0
#> 2 4      8      Difference    1.62     1.14       2.09 1                   0
#> 3 6      8      Difference    0.699    0.191      1.23 0.986               0
#>   prior.distribution prior.location prior.scale    bf10 method         
#>   <chr>                       <dbl>       <dbl>   <dbl> <chr>          
#> 1 cauchy                          0       0.707   11.4  Bayesian t-test
#> 2 cauchy                          0       0.707 5222.   Bayesian t-test
#> 3 cauchy                          0       0.707    5.36 Bayesian t-test
#>   log_e_bf10 label                          test.details    
#>        <dbl> <chr>                          <chr>           
#> 1       2.44 list(~log[e](BF['01'])==-2.44) Student's t-test
#> 2       8.56 list(~log[e](BF['01'])==-8.56) Student's t-test
#> 3       1.68 list(~log[e](BF['01'])==-1.68) Student's t-test

#------------------- within-subjects design ----------------------------

# parametric (Student's t-test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  subject.id = subject,
  type = "parametric",
  paired = TRUE,
  p.adjust.method = "BH"
)

#> # A tibble: 6 x 6
#>   group1 group2  p.value test.details     p.value.adjustment
#>   <chr>  <chr>     <dbl> <chr>            <chr>             
#> 1 HDHF   HDLF   1.06e- 3 Student's t-test FDR               
#> 2 HDHF   LDHF   7.02e- 2 Student's t-test FDR               
#> 3 HDHF   LDLF   3.95e-12 Student's t-test FDR               
#> 4 HDLF   LDHF   6.74e- 2 Student's t-test FDR               
#> 5 HDLF   LDLF   1.99e- 3 Student's t-test FDR               
#> 6 LDHF   LDLF   6.66e- 9 Student's t-test FDR               
#>   label                                    
#>   <chr>                                    
#> 1 list(~italic(p)[FDR-corrected]==0.001)   
#> 2 list(~italic(p)[FDR-corrected]==0.070)   
#> 3 list(~italic(p)[FDR-corrected]==3.95e-12)
#> 4 list(~italic(p)[FDR-corrected]==0.067)   
#> 5 list(~italic(p)[FDR-corrected]==0.002)   
#> 6 list(~italic(p)[FDR-corrected]==6.66e-09)

# non-parametric (Durbin-Conover test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  subject.id = subject,
  type = "nonparametric",
  paired = TRUE,
  p.adjust.method = "BY"
)

#> # A tibble: 6 x 11
#>   group1 group2 statistic  p.value alternative
#>   <chr>  <chr>      <dbl>    <dbl> <chr>      
#> 1 HDHF   HDLF        4.78 1.44e- 5 two.sided  
#> 2 HDHF   LDHF        2.44 4.47e- 2 two.sided  
#> 3 HDHF   LDLF        8.01 5.45e-13 two.sided  
#> 4 HDLF   LDHF        2.34 4.96e- 2 two.sided  
#> 5 HDLF   LDLF        3.23 5.05e- 3 two.sided  
#> 6 LDHF   LDLF        5.57 4.64e- 7 two.sided  
#>   method                                                                
#>   <chr>                                                                 
#> 1 Durbin's all-pairs test for a two-way balanced incomplete block design
#> 2 Durbin's all-pairs test for a two-way balanced incomplete block design
#> 3 Durbin's all-pairs test for a two-way balanced incomplete block design
#> 4 Durbin's all-pairs test for a two-way balanced incomplete block design
#> 5 Durbin's all-pairs test for a two-way balanced incomplete block design
#> 6 Durbin's all-pairs test for a two-way balanced incomplete block design
#>   distribution p.adjustment test.details        p.value.adjustment
#>   <chr>        <chr>        <chr>               <chr>             
#> 1 t            none         Durbin-Conover test BY                
#> 2 t            none         Durbin-Conover test BY                
#> 3 t            none         Durbin-Conover test BY                
#> 4 t            none         Durbin-Conover test BY                
#> 5 t            none         Durbin-Conover test BY                
#> 6 t            none         Durbin-Conover test BY                
#>   label                                   
#>   <chr>                                   
#> 1 list(~italic(p)[BY-corrected]==1.44e-05)
#> 2 list(~italic(p)[BY-corrected]==0.045)   
#> 3 list(~italic(p)[BY-corrected]==5.45e-13)
#> 4 list(~italic(p)[BY-corrected]==0.050)   
#> 5 list(~italic(p)[BY-corrected]==0.005)   
#> 6 list(~italic(p)[BY-corrected]==4.64e-07)

# robust (Yuen's trimmed means t-test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  subject.id = subject,
  type = "robust",
  paired = TRUE,
  p.adjust.method = "hommel"
)

#> # A tibble: 6 x 10
#>   group1 group2 estimate conf.low conf.high  p.value  p.crit
#>   <chr>  <chr>     <dbl>    <dbl>     <dbl>    <dbl>   <dbl>
#> 1 HDHF   HDLF      1.16    0.318      2.00  1.49e- 3 0.0127 
#> 2 HDHF   LDHF      0.5    -0.188      1.19  6.20e- 2 0.025  
#> 3 HDHF   LDLF      2.10    1.37       2.82  1.79e-10 0.00851
#> 4 HDLF   LDHF     -0.701  -1.71       0.303 6.20e- 2 0.05   
#> 5 HDLF   LDLF      0.938   0.0694     1.81  1.36e- 2 0.0169 
#> 6 LDHF   LDLF      1.54    0.810      2.27  1.16e- 6 0.0102 
#>   test.details              p.value.adjustment
#>   <chr>                     <chr>             
#> 1 Yuen's trimmed means test Hommel            
#> 2 Yuen's trimmed means test Hommel            
#> 3 Yuen's trimmed means test Hommel            
#> 4 Yuen's trimmed means test Hommel            
#> 5 Yuen's trimmed means test Hommel            
#> 6 Yuen's trimmed means test Hommel            
#>   label                                       
#>   <chr>                                       
#> 1 list(~italic(p)[Hommel-corrected]==0.001)   
#> 2 list(~italic(p)[Hommel-corrected]==0.062)   
#> 3 list(~italic(p)[Hommel-corrected]==1.79e-10)
#> 4 list(~italic(p)[Hommel-corrected]==0.062)   
#> 5 list(~italic(p)[Hommel-corrected]==0.014)   
#> 6 list(~italic(p)[Hommel-corrected]==1.16e-06)

# Bayes Factor (Student's t-test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  subject.id = subject,
  type = "bayes",
  paired = TRUE
)

#> # A tibble: 6 x 16
#>   group1 group2 term       estimate conf.low conf.high    pd rope.percentage
#>   <chr>  <chr>  <chr>         <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 HDHF   HDLF   Difference   -1.10    -1.62    -0.591  1.00            0    
#> 2 HDHF   LDHF   Difference   -0.454   -0.814   -0.0130 0.966           0.183
#> 3 HDHF   LDLF   Difference   -2.13    -2.58    -1.75   1               0    
#> 4 HDLF   LDHF   Difference    0.650    0.103    1.20   0.973           0.133
#> 5 HDLF   LDLF   Difference   -0.975   -1.45    -0.493  1.00            0    
#> 6 LDHF   LDLF   Difference   -1.66    -2.08    -1.28   1               0    
#>   prior.distribution prior.location prior.scale     bf10 method         
#>   <chr>                       <dbl>       <dbl>    <dbl> <chr>          
#> 1 cauchy                          0       0.707 4.16e+ 1 Bayesian t-test
#> 2 cauchy                          0       0.707 5.83e- 1 Bayesian t-test
#> 3 cauchy                          0       0.707 1.20e+10 Bayesian t-test
#> 4 cauchy                          0       0.707 6.98e- 1 Bayesian t-test
#> 5 cauchy                          0       0.707 1.81e+ 1 Bayesian t-test
#> 6 cauchy                          0       0.707 4.81e+ 6 Bayesian t-test
#>   log_e_bf10 label                           test.details    
#>        <dbl> <chr>                           <chr>           
#> 1      3.73  list(~log[e](BF['01'])==-3.73)  Student's t-test
#> 2     -0.539 list(~log[e](BF['01'])==0.54)   Student's t-test
#> 3     23.2   list(~log[e](BF['01'])==-23.21) Student's t-test
#> 4     -0.359 list(~log[e](BF['01'])==0.36)   Student's t-test
#> 5      2.90  list(~log[e](BF['01'])==-2.90)  Student's t-test
#> 6     15.4   list(~log[e](BF['01'])==-15.39) Student's t-test

# }