Multiple pairwise comparison tests with tidy data

Calculate parametric, non-parametric, robust, and Bayes Factor pairwise comparisons between group levels with corrections for multiple testing.

pairwise_comparisons(
  data,
  x,
  y,
  type = "parametric",
  paired = FALSE,
  var.equal = FALSE,
  tr = 0.1,
  bf.prior = 0.707,
  p.adjust.method = "holm",
  k = 2L,
  ...
)

pairwise_p(
  data,
  x,
  y,
  type = "parametric",
  paired = FALSE,
  var.equal = FALSE,
  tr = 0.1,
  bf.prior = 0.707,
  p.adjust.method = "holm",
  k = 2L,
  ...
)

Arguments

data
A dataframe from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

type
Type of statistic expected ("parametric" or "nonparametric" or "robust" or "bayes").Corresponding abbreviations are also accepted: "p" (for parametric), "np" (nonparametric), "r" (robust), or "bf"resp.

paired
Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

var.equal
a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

tr
Trim level for the mean when carrying out robust tests. If you get error stating "Standard error cannot be computed because of Winsorized variance of 0 (e.g., due to ties). Try to decrease the trimming level.", try to play around with the value of tr, which is by default set to 0.1. Lowering the value might help.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

p.adjust.method
Adjustment method for p-values for multiple comparisons. Possible methods are: "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".

k
Number of digits after decimal point (should be an integer) (Default: k = 2L).

...
Current ignored.

Value

A tibble dataframe containing two columns corresponding to group levels being compared with each other (group1 and group2) and p.value column corresponding to this comparison. The dataframe will also contain a p.value.label column containing a label for this p-value, in case this needs to be displayed in geom_ggsignif. In addition to these common columns across the different types of statistics, there will be additional columns specific to the type of test being run.

This function provides a unified syntax to carry out pairwise comparison tests and internally relies on other packages to carry out these tests. For more details about the included tests, see the documentation for the respective functions:

parametric : stats::pairwise.t.test() (paired) and PMCMRplus::gamesHowellTest() (unpaired)

non-parametric : PMCMRplus::durbinAllPairsTest() (paired) and PMCMRplus::kwAllPairsDunnTest() (unpaired)

robust : WRS2::rmmcp() (paired) and WRS2::lincon() (unpaired)

Bayes Factor : BayesFactor::ttestBF()

The significance column asterisks indicate significance levels of p-values in the American Psychological Association (APA) mandated format:

ns : > 0.05

* : < 0.05

** : < 0.01

*** : < 0.001

Examples

# \donttest{
# for reproducibility
set.seed(123)
library(pairwiseComparisons)

# show me all columns and make the column titles bold
options(tibble.width = Inf, pillar.bold = TRUE)

#------------------- between-subjects design ----------------------------

# parametric
# if `var.equal = TRUE`, then Student's t-test will be run
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "parametric",
  var.equal = TRUE,
  paired = FALSE,
  p.adjust.method = "none"
)

#> # A tibble: 3 x 7
#>   group1 group2     p.value significance label                               
#>   <chr>  <chr>        <dbl> <chr>        <chr>                               
#> 1 4      6      0.0106      *            list(~italic(p)[uncorrected]==0.011)
#> 2 4      8      0.000000207 ***          list(~italic(p)[uncorrected]<=0.001)
#> 3 6      8      0.00516     **           list(~italic(p)[uncorrected]==0.005)
#>   test.details     p.value.adjustment
#>   <chr>            <chr>             
#> 1 Student's t-test None              
#> 2 Student's t-test None              
#> 3 Student's t-test None              

# if `var.equal = FALSE`, then Games-Howell test will be run
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "parametric",
  var.equal = FALSE,
  paired = FALSE,
  p.adjust.method = "bonferroni"
)

#> # A tibble: 3 x 8
#>   group1 group2 statistic   p.value significance
#>   <chr>  <chr>      <dbl>     <dbl> <chr>       
#> 1 4      6           5.39 0.0125    *           
#> 2 4      8           9.11 0.0000124 ***         
#> 3 6      8           5.12 0.0148    *           
#>   label                                         test.details     
#>   <chr>                                         <chr>            
#> 1 list(~italic(p)[Bonferroni-corrected]==0.012) Games-Howell test
#> 2 list(~italic(p)[Bonferroni-corrected]<=0.001) Games-Howell test
#> 3 list(~italic(p)[Bonferroni-corrected]==0.015) Games-Howell test
#>   p.value.adjustment
#>   <chr>             
#> 1 Bonferroni        
#> 2 Bonferroni        
#> 3 Bonferroni        

# non-parametric (Dunn test)
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "nonparametric",
  paired = FALSE,
  p.adjust.method = "none"
)

#> # A tibble: 3 x 8
#>   group1 group2 statistic    p.value significance
#>   <chr>  <chr>      <dbl>      <dbl> <chr>       
#> 1 4      6           1.84 0.0663     ns          
#> 2 4      8           4.76 0.00000198 ***         
#> 3 6      8           2.22 0.0263     *           
#>   label                                test.details p.value.adjustment
#>   <chr>                                <chr>        <chr>             
#> 1 list(~italic(p)[uncorrected]==0.066) Dunn test    None              
#> 2 list(~italic(p)[uncorrected]<=0.001) Dunn test    None              
#> 3 list(~italic(p)[uncorrected]==0.026) Dunn test    None              

# robust (Yuen's trimmed means t-test)
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "robust",
  paired = FALSE,
  p.adjust.method = "fdr"
)

#> # A tibble: 3 x 10
#>   group1 group2 estimate conf.low conf.high  p.value significance
#>   <chr>  <chr>     <dbl>    <dbl>     <dbl>    <dbl> <chr>       
#> 1 4      6        -0.846    -1.51    -0.180 0.00625  **          
#> 2 4      8        -1.68     -2.48    -0.878 0.000258 ***         
#> 3 6      8        -0.832    -1.53    -0.131 0.00625  **          
#>   label                                  test.details             
#>   <chr>                                  <chr>                    
#> 1 list(~italic(p)[FDR-corrected]==0.006) Yuen's trimmed means test
#> 2 list(~italic(p)[FDR-corrected]<=0.001) Yuen's trimmed means test
#> 3 list(~italic(p)[FDR-corrected]==0.006) Yuen's trimmed means test
#>   p.value.adjustment
#>   <chr>             
#> 1 FDR               
#> 2 FDR               
#> 3 FDR               

# Bayes Factor (Student's t-test)
pairwise_comparisons(
  data = mtcars,
  x = cyl,
  y = wt,
  type = "bayes",
  paired = FALSE
)

#> # A tibble: 3 x 17
#>   group1 group2 term       estimate conf.low conf.high    pd rope.percentage
#>   <chr>  <chr>  <chr>         <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 4      6      Difference    0.688    0.252      1.10 0.994               0
#> 2 4      8      Difference    1.62     1.15       2.09 1                   0
#> 3 6      8      Difference    0.703    0.207      1.24 0.988               0
#>   prior.distribution prior.location prior.scale effects component      bf10
#>   <chr>                       <dbl>       <dbl> <chr>   <chr>         <dbl>
#> 1 cauchy                          0       0.707 fixed   conditional   11.4 
#> 2 cauchy                          0       0.707 fixed   conditional 5222.  
#> 3 cauchy                          0       0.707 fixed   conditional    5.36
#>   log_e_bf10 label                       test.details    
#>        <dbl> <chr>                       <chr>           
#> 1       2.44 list(~log[e](BF[10])==2.44) Student's t-test
#> 2       8.56 list(~log[e](BF[10])==8.56) Student's t-test
#> 3       1.68 list(~log[e](BF[10])==1.68) Student's t-test

#------------------- within-subjects design ----------------------------

# parametric (Student's t-test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  type = "parametric",
  paired = TRUE,
  p.adjust.method = "BH"
)

#> # A tibble: 6 x 7
#>   group1 group2  p.value significance label                                 
#>   <chr>  <chr>     <dbl> <chr>        <chr>                                 
#> 1 HDHF   HDLF   1.06e- 3 **           list(~italic(p)[FDR-corrected]==0.001)
#> 2 HDHF   LDHF   7.02e- 2 ns           list(~italic(p)[FDR-corrected]==0.070)
#> 3 HDHF   LDLF   3.95e-12 ***          list(~italic(p)[FDR-corrected]<=0.001)
#> 4 HDLF   LDHF   6.74e- 2 ns           list(~italic(p)[FDR-corrected]==0.067)
#> 5 HDLF   LDLF   1.99e- 3 **           list(~italic(p)[FDR-corrected]==0.002)
#> 6 LDHF   LDLF   6.66e- 9 ***          list(~italic(p)[FDR-corrected]<=0.001)
#>   test.details     p.value.adjustment
#>   <chr>            <chr>             
#> 1 Student's t-test FDR               
#> 2 Student's t-test FDR               
#> 3 Student's t-test FDR               
#> 4 Student's t-test FDR               
#> 5 Student's t-test FDR               
#> 6 Student's t-test FDR               

# non-parametric (Durbin-Conover test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  type = "nonparametric",
  paired = TRUE,
  p.adjust.method = "BY"
)

#> # A tibble: 6 x 8
#>   group1 group2 statistic  p.value significance
#>   <chr>  <chr>      <dbl>    <dbl> <chr>       
#> 1 HDHF   HDLF        4.78 1.44e- 5 ***         
#> 2 HDHF   LDHF        2.44 4.47e- 2 *           
#> 3 HDHF   LDLF        8.01 5.45e-13 ***         
#> 4 HDLF   LDHF        2.34 4.96e- 2 *           
#> 5 HDLF   LDLF        3.23 5.05e- 3 **          
#> 6 LDHF   LDLF        5.57 4.64e- 7 ***         
#>   label                                 test.details        p.value.adjustment
#>   <chr>                                 <chr>               <chr>             
#> 1 list(~italic(p)[BY-corrected]<=0.001) Durbin-Conover test BY                
#> 2 list(~italic(p)[BY-corrected]==0.045) Durbin-Conover test BY                
#> 3 list(~italic(p)[BY-corrected]<=0.001) Durbin-Conover test BY                
#> 4 list(~italic(p)[BY-corrected]==0.050) Durbin-Conover test BY                
#> 5 list(~italic(p)[BY-corrected]==0.005) Durbin-Conover test BY                
#> 6 list(~italic(p)[BY-corrected]<=0.001) Durbin-Conover test BY                

# robust (Yuen's trimmed means t-test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  type = "robust",
  paired = TRUE,
  p.adjust.method = "hommel"
)

#> # A tibble: 6 x 10
#>   group1 group2 estimate conf.low conf.high  p.value significance
#>   <chr>  <chr>     <dbl>    <dbl>     <dbl>    <dbl> <chr>       
#> 1 HDHF   HDLF      1.16    0.318      2.00  1.49e- 3 **          
#> 2 HDHF   LDHF      0.5    -0.188      1.19  6.20e- 2 ns          
#> 3 HDHF   LDLF      2.10    1.37       2.82  1.79e-10 ***         
#> 4 HDLF   LDHF     -0.701  -1.71       0.303 6.20e- 2 ns          
#> 5 HDLF   LDLF      0.938   0.0694     1.81  1.36e- 2 *           
#> 6 LDHF   LDLF      1.54    0.810      2.27  1.16e- 6 ***         
#>   label                                     test.details             
#>   <chr>                                     <chr>                    
#> 1 list(~italic(p)[Hommel-corrected]==0.001) Yuen's trimmed means test
#> 2 list(~italic(p)[Hommel-corrected]==0.062) Yuen's trimmed means test
#> 3 list(~italic(p)[Hommel-corrected]<=0.001) Yuen's trimmed means test
#> 4 list(~italic(p)[Hommel-corrected]==0.062) Yuen's trimmed means test
#> 5 list(~italic(p)[Hommel-corrected]==0.014) Yuen's trimmed means test
#> 6 list(~italic(p)[Hommel-corrected]<=0.001) Yuen's trimmed means test
#>   p.value.adjustment
#>   <chr>             
#> 1 Hommel            
#> 2 Hommel            
#> 3 Hommel            
#> 4 Hommel            
#> 5 Hommel            
#> 6 Hommel            

# Bayes Factor (Student's t-test)
pairwise_comparisons(
  data = bugs_long,
  x = condition,
  y = desire,
  type = "bayes",
  paired = TRUE
)

#> # A tibble: 6 x 17
#>   group1 group2 term       estimate conf.low conf.high    pd rope.percentage
#>   <chr>  <chr>  <chr>         <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 HDHF   HDLF   Difference   -1.11    -1.67    -0.635  1             0      
#> 2 HDHF   LDHF   Difference   -0.445   -0.864   -0.0589 0.966         0.0320 
#> 3 HDHF   LDLF   Difference   -2.13    -2.54    -1.72   1             0      
#> 4 HDLF   LDHF   Difference    0.657    0.142    1.23   0.974         0.00112
#> 5 HDLF   LDLF   Difference   -0.981   -1.46    -0.499  0.999         0      
#> 6 LDHF   LDLF   Difference   -1.65    -2.06    -1.23   1             0      
#>   prior.distribution prior.location prior.scale effects component       bf10
#>   <chr>                       <dbl>       <dbl> <chr>   <chr>          <dbl>
#> 1 cauchy                          0       0.707 fixed   conditional 4.16e+ 1
#> 2 cauchy                          0       0.707 fixed   conditional 5.83e- 1
#> 3 cauchy                          0       0.707 fixed   conditional 1.20e+10
#> 4 cauchy                          0       0.707 fixed   conditional 6.98e- 1
#> 5 cauchy                          0       0.707 fixed   conditional 1.81e+ 1
#> 6 cauchy                          0       0.707 fixed   conditional 4.81e+ 6
#>   log_e_bf10 label                        test.details    
#>        <dbl> <chr>                        <chr>           
#> 1      3.73  list(~log[e](BF[10])==3.73)  Student's t-test
#> 2     -0.539 list(~log[e](BF[10])==-0.54) Student's t-test
#> 3     23.2   list(~log[e](BF[10])==23.21) Student's t-test
#> 4     -0.359 list(~log[e](BF[10])==-0.36) Student's t-test
#> 5      2.90  list(~log[e](BF[10])==2.90)  Student's t-test
#> 6     15.4   list(~log[e](BF[10])==15.39) Student's t-test

# }